refactor: one test file per period
